fix: widen stream-count columns from int32 to int64
THE BUG (root cause of the entire session). Every popular album was
failing the AlbumCache UPDATE with asyncpg.DataError:

    invalid input for query argument $1: 2174635255 (value out of int32 range)
    SQL: UPDATE album_cache SET kworb_streams=$1::INTEGER ...

kworb_streams was declared sa.Integer, which is INT4 on Postgres
(max 2,147,483,647). UTOPIA streams ~7B, Astroworld ~11B, folklore,
Donda — all overflow. asyncpg rejected the parameter, transaction
rolled back, row stayed "pending", no exception bubbled out of
_enrich_album's try/except wrapper in a way that surfaced.

The bug was masked the entire session — closed sessions, FastAPI
BackgroundTasks not dispatching, asyncio GC, synchronous /streams,
market=US — were all wrong theories. The only thing that surfaced
the real error was instrumenting save_kworb_streams to count branch
outcomes and capture the asyncpg exception text in last_call_outcome.
save_stats showed: 40 called / 40 row_found / 0 committed / 40
commit_failed — every single save attempt was crashing inside the
commit on the int32 overflow.

Fix:
  models.AlbumCache.kworb_streams:    Integer -> BigInteger (int8)
  models.StreamAnchor.stream_count:   Integer -> BigInteger (int8)
  Alembic migration u1v2w3x4y5z6 ALTERs both columns on Postgres.
    Existing data fits losslessly (int32 ⊂ int64), so the migration
    is fast and non-destructive.

Regression test added (test_enrich_album_handles_streams_above_int32_max):
asserts an explicit 7,068,208,275 stream count (real UTOPIA value)
persists end-to-end through _enrich_album. Would have caught this on
day one. 33/33 pass.

Once Railway deploys this and runs the migration, the existing 706
pending rows will start transitioning correctly on next sweeper cycle.
UTOPIA, folklore, every popular album finally lands.

+++ backend/models.py
@@ -3,7 +3,7 @@
 from datetime import datetime
 from typing import Optional
 
-from sqlalchemy import DateTime, Float, Integer, String, Text
+from sqlalchemy import BigInteger, DateTime, Float, Integer, String, Text
 from sqlalchemy.orm import Mapped, mapped_column
 
 from database import Base
@@ -126,7 +126,10 @@ class StreamAnchor(Base):
     entity_id: Mapped[str] = mapped_column(String(64), index=True)
     entity_type: Mapped[str] = mapped_column(String(16))   # "track" | "album"
     snapshot_date: Mapped[str] = mapped_column(String(16)) # ISO date "YYYY-MM-DD"
-    stream_count: Mapped[int] = mapped_column(Integer)
+    # BigInteger (int64) — Postgres INT4 caps at 2.1B and popular albums
+    # routinely exceed that (UTOPIA ~7B, Astroworld ~11B). The same overflow
+    # would happen here as on AlbumCache.kworb_streams.
+    stream_count: Mapped[int] = mapped_column(BigInteger)
     source: Mapped[str] = mapped_column(String(32))        # "kworb_daily" | "wayback"
     fetched_at: Mapped[datetime] = mapped_column(DateTime, default=datetime.utcnow)
 
@@ -239,8 +242,12 @@ class AlbumCache(Base):
     popularity: Mapped[Optional[int]] = mapped_column(Integer, nullable=True)
     image_url: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
 
-    # Stream enrichment — filled in async by Kworb scrape
-    kworb_streams: Mapped[Optional[int]] = mapped_column(Integer, nullable=True)
+    # Stream enrichment — filled in async by Kworb scrape.
+    # BigInteger (int64): Postgres INT4 maxes at 2.1B and popular albums
+    # routinely exceed that. UTOPIA ~7B, Astroworld ~11B. Using Integer
+    # here would silently break enrichment for every popular album with
+    # asyncpg DataError: "value out of int32 range" on the UPDATE.
+    kworb_streams: Mapped[Optional[int]] = mapped_column(BigInteger, nullable=True)
     # "pending" | "done" | "failed"
     enrichment_status: Mapped[str] = mapped_column(String(16), default="pending")
     enriched_at: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
